Add an extractor for app_id
This is generally useful thing to be able to easily get out of a
request and even thought it is possible to do currently (via the
delivery) it is more cumbersome than it has to be. This PR makes it
straight forward by having a simple wrapper that implements the
`Extractor` trait.

    use kanin::extract::AppId;
    async fn my_handler(AppId(app_id): AppId) {
        dbg!(app_id);
    }